Cleaned up project
I was building with a sucky version of the android sdk. So it created an
annoying little library when I imported the project (more info:
https://code.google.com/p/android/issues/detail?id=67513 ). So I removed
that library from the dependencies (explicitly shown in
project.properties) and cleaned up the project so that it wouldn't need
to run based on that library. I also cleaned up SoundPlayer's action bar
and allowed the .classpath and .project to be committed, since they
weren't auto-generated on my computer and caused lots of heartache. If
anyone else is using that version (i.e. if you see an app_compat_v7 in
your package explorer), here's the page I used to remove the library:
http://stackoverflow.com/questions/22769016/how-to-remove-support-library-appcompat-v7-eclipse